% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_page_numbers.R
\name{add_filetype}
\alias{add_filetype}
\title{Adds a filetype column to the readtext object!}
\usage{
add_filetype(readtext_object)
}
\arguments{
\item{readtext_object}{a readtext s3 class created by load_documents}
}
\value{
readtext object with new filetype column
}
\description{
Adds a filetype column to the readtext object!
}
