% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_documents.R
\name{convert_to_corpus}
\alias{convert_to_corpus}
\title{Convert to corpus for easy searching}
\usage{
convert_to_corpus(readtext_object)
}
\arguments{
\item{readtext_object}{a readtext s3 class created by load_documents}
}
\value{
quanteda corpus
}
\description{
Convert to corpus for easy searching
}
