% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_documents.R
\name{convert_to_tokens}
\alias{convert_to_tokens}
\title{Convert to tokens for easy searching}
\usage{
convert_to_tokens(readtext_object)
}
\arguments{
\item{readtext_object}{a readtext s3 class created by load_documents}
}
\value{
quanteda tokenized corpus
}
\description{
Convert to tokens for easy searching
}
