% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_documents.R
\name{load_documents}
\alias{load_documents}
\title{A test function}
\usage{
load_documents(path)
}
\arguments{
\item{path}{character vector with the path for the location of document(s)}
}
\value{
readtext object(s) of the loaded files
}
\description{
A test function
}
\examples{
load_documents
}
