% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_documents.R
\name{load_documents_recursive}
\alias{load_documents_recursive}
\title{Loading all supported documents in folders and subfolders}
\usage{
load_documents_recursive(path)
}
\arguments{
\item{path}{\itemize{
\item path to the folder that contains all documents
}}
}
\value{
the loaded readtext object already with page added
}
\description{
Loading all supported documents in folders and subfolders
}
\examples{
# not run
# readtext_object <- load_documents_recursive()
}
