% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_documents.R
\name{main}
\alias{main}
\title{main function to interface with docuSearchR}
\usage{
main(path, list_of_regexes, window = 10)
}
\arguments{
\item{path}{path to the folder that contains documents - also reads files in nested folders}

\item{list_of_regexes}{list of searchterms - regexcompatible}

\item{window}{amount of context-words to get around the matches!}
}
\value{
the simplified search results!
}
\description{
main function to interface with docuSearchR
}
\examples{
# not run
# main("path to folder", c("regex"), window = 5)
}
