% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_documents.R
\name{read_excel}
\alias{read_excel}
\title{function to read in a single excel sheet}
\usage{
read_excel(path)
}
\arguments{
\item{path}{path to the excel file!}
}
\value{
readtextobject with the charactercells from the sheet!
}
\description{
function to read in a single excel sheet
}
