% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searching_kwic.R
\name{search_for_kwic}
\alias{search_for_kwic}
\title{Using quantedas kwic to search for the keyword in context!}
\usage{
search_for_kwic(tokens, pattern, ...)
}
\arguments{
\item{tokens}{tokens object created with convert_to_corpus}

\item{pattern}{regex to search for}

\item{...}{additional parameters passed to kwic!}
}
\description{
Using quantedas kwic to search for the keyword in context!
}
