% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searching_kwic.R
\name{search_for_kws}
\alias{search_for_kws}
\title{Search for multiple patterns}
\usage{
search_for_kws(tokens, patterns, window = 10, ...)
}
\arguments{
\item{tokens}{\itemize{
\item quanteda token object - created with convert_to_quanteda
}}

\item{patterns}{\itemize{
\item list of regexes to search for
}}

\item{window}{\itemize{
\item amount of surrounding tokens to get as context
}}

\item{...}{additional parameters passed to quanteda::kwic}
}
\value{
kwic match dataframe
}
\description{
Search for multiple patterns
}
