% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searching_kwic.R
\name{simplify_kwic}
\alias{simplify_kwic}
\title{simplify kwic dataframe for the end user}
\usage{
simplify_kwic(kwic)
}
\arguments{
\item{kwic}{\itemize{
\item kwic dataframe created by search_for_kws
}}
}
\description{
simplify kwic dataframe for the end user
}
