% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_documents.R
\name{load_documents}
\alias{load_documents}
\title{load_documents - loading word and pdf files for easy reading!}
\usage{
load_documents(path)
}
\arguments{
\item{path}{character vector with the path for the location of documents}
}
\value{
readtext objects of the loaded files
}
\description{
load_documents - loading word and pdf files for easy reading!
}
\examples{
load_documents
}
